+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/EdcDataPlaneClient.java
@@ -32,6 +32,7 @@ import org.eclipse.edc.spi.types.domain.edr.EndpointDataReference;
 import org.eclipse.tractusx.irs.data.StringMapper;
 import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
 import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotificationResponse;
+import org.eclipse.tractusx.irs.edc.client.model.notification.NotificationContent;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -64,14 +65,6 @@ public class EdcDataPlaneClient {
         return extractData(response);
     }
 
-    private String getUrl(final String connectorUrl, final String subUrl) {
-        var url = connectorUrl;
-        if (subUrl != null && !subUrl.isEmpty()) {
-            url = url.endsWith("/") ? url + subUrl : url + "/" + subUrl;
-        }
-        return url;
-    }
-
     private HttpHeaders headers(final EndpointDataReference dataReference) {
         final HttpHeaders headers = new HttpHeaders();
         headers.setAccept(List.of(MediaType.APPLICATION_JSON));
@@ -93,9 +86,9 @@ public class EdcDataPlaneClient {
         return modifiedResponse;
     }
 
-    public EdcNotificationResponse sendData(final EndpointDataReference dataReference, final String subUrl,
-            final EdcNotification notification) {
-        final String url = getUrl(dataReference.getEndpoint(), subUrl);
+    public EdcNotificationResponse sendData(final EndpointDataReference dataReference,
+            final EdcNotification<NotificationContent> notification) {
+        final String url = dataReference.getEndpoint();
         final HttpHeaders headers = headers(dataReference);
         headers.setContentType(MediaType.APPLICATION_JSON);