+++ irs-edc-client/src/test/java/org/eclipse/tractusx/irs/edc/client/EdcSubmodelClientTest.java
@@ -28,6 +28,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.io.IOException;
@@ -66,6 +68,7 @@ import org.eclipse.tractusx.irs.edc.client.model.CatalogItem;
 import org.eclipse.tractusx.irs.edc.client.model.NegotiationResponse;
 import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
 import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotificationResponse;
+import org.eclipse.tractusx.irs.edc.client.model.notification.NotificationContent;
 import org.eclipse.tractusx.irs.testing.containers.LocalTestDataConfigurationAware;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
@@ -114,7 +117,6 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         config.getControlplane().setProviderSuffix(PROVIDER_SUFFIX);
 
         config.setSubmodel(new EdcConfiguration.SubmodelConfig());
-        config.getSubmodel().setPath("/submodel");
         config.getSubmodel().setUrnPrefix("/urn");
         config.getSubmodel().setRequestTtl(Duration.ofMinutes(10));
         testee = new EdcSubmodelClientImpl(config, contractNegotiationService, edcDataPlaneClient,
@@ -144,21 +146,22 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
     @Test
     void shouldSendNotificationSuccessfully() throws Exception {
         // arrange
-        final EdcNotification notification = EdcNotification.builder().build();
-        when(catalogFacade.fetchCatalogById(any(), any())).thenReturn(
+        final EdcNotification<NotificationContent> notification = EdcNotification.builder().build();
+        when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId("itemId").build()));
         when(contractNegotiationService.negotiate(any(), any())).thenReturn(
                 NegotiationResponse.builder().contractAgreementId("agreementId").build());
         final EndpointDataReference ref = mock(EndpointDataReference.class);
         endpointDataReferenceStorage.put("agreementId", ref);
-        when(edcDataPlaneClient.sendData(eq(ref), any(), eq(notification))).thenReturn(() -> true);
+        when(edcDataPlaneClient.sendData(ref, notification)).thenReturn(() -> true);
 
         // act
-        final var result = testee.sendNotification(ENDPOINT_ADDRESS, "notify-request-asset", notification);
+        final var result = testee.sendNotification(CONNECTOR_ENDPOINT, "notify-request-asset", notification);
         final EdcNotificationResponse response = result.get(5, TimeUnit.SECONDS);
 
         // assert
         assertThat(response.deliveredSuccessfully()).isTrue();
+        verify(contractNegotiationService, times(1)).negotiate(eq(CONNECTOR_ENDPOINT + PROVIDER_SUFFIX), any());
     }
 
     @Test