+++ local/testing/api-tests/tavern_helpers.py
@@ -5,7 +5,7 @@ import requests
 import os
 
 
-def supplyChainImpacted_is_correct_in_submodels_for_valid_ID(response):
+def supplyChainImpacted_is_Yes(response):
     submodels = response.json().get("submodels")
     print("submodels ", submodels)
     assert len(submodels) <= 1
@@ -14,7 +14,7 @@ def supplyChainImpacted_is_correct_in_submodels_for_valid_ID(response):
         assert 'Yes' in i.get("payload").get('supplyChainImpacted')
 
 
-def supplyChainImpacted_is_correct_in_submodels_for_valid_ID_in_unknown_BPN(response):
+def supplyChainImpacted_is_No(response):
     submodels = response.json().get("submodels")
     print("submodels ", submodels)
     assert len(submodels) <= 1
@@ -23,7 +23,7 @@ def supplyChainImpacted_is_correct_in_submodels_for_valid_ID_in_unknown_BPN(resp
         assert 'No' in i.get("payload").get('supplyChainImpacted')
 
 
-def supplyChainImpacted_is_correct_in_submodels_for_unknown_ID(response):
+def supplyChainImpacted_is_Unknown(response):
     submodels = response.json().get("submodels")
     print("submodels ", submodels)
     assert len(submodels) <= 1
@@ -35,15 +35,30 @@ def supplyChainImpacted_is_correct_in_submodels_for_unknown_ID(response):
 def errors_for_invalid_investigation_request_are_correct(response):
     print(response.json().get("messages"))
     error_list = response.json().get("messages")
-    assert 'incidentBpns:must not be empty' in error_list
-    assert 'globalAssetId:must not be blank' in error_list
+    assert 'incidentBPNSs:must not be empty' in error_list
+    assert 'key.globalAssetId:must not be blank' in error_list
+    assert 'key.bpn:must not be blank' in error_list
 
 
 def relationships_for_BPN_investigations_contains_several_childs(response):
     relationships = response.json().get("relationships")
     print("relationships: ", relationships)
     print("Länge: ", len(relationships))
-    assert len(relationships) > 1
+    assert len(relationships) != 0
+
+def ESS_job_parameter_are_as_requested(response):
+    print("Check if ESS-job parameter are as requested:")
+    parameter = response.json().get('job').get('parameter')
+    print(parameter)
+    assert parameter.get('bomLifecycle') == 'asPlanned'
+    assert parameter.get('collectAspects') is True
+    assert parameter.get('depth') == 1
+    assert parameter.get('direction') == 'downward'
+    assert parameter.get('lookupBPNs') is False
+    #assert parameter.get('callbackUrl') == 'https://www.check123.com'
+    aspects_list = parameter.get("aspects")
+    assert 'PartSiteInformationAsPlanned' in aspects_list
+    assert 'PartAsPlanned' in aspects_list
 
 
 ############################## /\ ESS helpers /\ ##############################