+++ charts/irs-helm/CHANGELOG.md
@@ -6,10 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.7.1] - 2023-09-29
+### Changed
+- Added toString template for `edc.controlplane.apikey.secret`
+
+## [6.7.0] - 2023-09-28
+### Changed
+- Update IRS version to 3.5.0
+
+### Removed
+- Removed unused configuration parameters `edc.submodel.path`,`discovery.endpoint`, `discovery.mockEdcAddres`
+
 ## [6.6.1] - 2023-09-22
 ### Changed
 - Update IRS version to 3.4.1
-- 
+
 ## [6.6.0] - 2023-09-01
 ### Changed
 - Update IRS version to 3.4.0